examples/usecases/bulk_upload/pyproject.toml
[tool.poetry]
name = "bulk_upload"
version = "0.1.0"
description = "Bulk upload services to Spark"
authors = ["Coherent <team@coherent.global>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.11"
cspark = "^0.2.3"
pyyaml = "^6.0.0"
jsonschema = "^4.24.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
